preVersion 5.0 (Added control from telegram)
